arl-backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.13.0
annotated-types==0.7.0
anyio==3.7.1
arq==0.25.0
async-timeout==5.0.1
asyncpg==0.29.0
attrs==25.4.0
bcrypt==4.0.1
bidict==0.23.1
black==23.12.0
certifi==2025.11.12
cffi==2.0.0
click==8.3.1
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.1.0
fastapi==0.104.1
frozenlist==1.8.0
greenlet==3.2.4
h11==0.16.0
hiredis==3.3.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.25.2
idna==3.11
iniconfig==2.3.0
isort==5.13.2
Mako==1.3.10
MarkupSafe==3.0.3
multidict==6.7.0
mypy==1.7.1
mypy_extensions==1.1.0
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.5.0
pluggy==1.6.0
propcache==0.4.1
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.23
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.2.1
python-engineio==4.12.3
python-jose==3.3.0
python-multipart==0.0.6
python-socketio==5.10.0
PyYAML==6.0.3
redis==5.0.1
rsa==4.9.1
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.15.0
uvicorn==0.24.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
wsproto==1.3.2
yarl==1.22.0
